LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/img.complete-expected.txt

   

PASS img src omitted 
PASS img src empty 
FAIL async src complete test assert_equals: expected "/3.jpg" but got "/http://localhost:8800/html/semantics/embedded-content/the-img-element/3.jpg"
FAIL IDL attribute complete returns true when image resource has been fetched but not run yet & image is not in broken state assert_unreached: .complete didn't change to true Reached unreachable code
PASS async src broken test 
